qemu: Fix qemu package generation
The PACKAGE_QEMU_TRUSTY option has not been used in a while.
This fixes the output files so that an output package now
runs. Add an additional PACKAGE_QEMU_WITHOUT_ANDROID option
that excludes the AOSP prebuilts, in case someone wants to use
the qemu package with their own build of Android.

Bug: 230135749
Test: build package, extract to directory, execute run script
Change-Id: I949cf90eba163ba2a736bbcb51a71a282ebd7ba6