Use TelephonyManager.getNetworkCurrentIso() for country code.
When there are no Telephony manager subscriptions found, the current
code stops looking for other country code sources. We change this
behavior to make a call to TelephonyManager.getNetworkCurrentIso(), in
case it has a valid country code available. The code will also continue
to attempt to get a location from the Fused Provider.

Test: atest ServiceUwbTests
Test: Manually tested on phone with no SIM and GPS disabled. In this
case, getNetworkCurrentIso() also returned an empty string.
Bug: 299961333

Change-Id: Ife5da0116feb4cec126b625df2b6629546f017fd