qemu: Add PCI bridge and move virtio console devices
On qemu/x86 we ran out of PCI slots which makes it impossible to add
more virtio devices and we are also close to running out of PCI slots
on qemu/arm64.

In contrast crosvm has 12 more PCI slots available because it uses a
PCI bridge.

Add a PCI bridge for qemu so that we can extend the number of PCI
slots and to support adding more virtio devices.

Also move all virtio console devices to the newly added bus since qemu
always uses the first PCI bus one when not specified.

Bug: 307608726
Test: boot qemu x86/arm64/riscv64
Change-Id: I9c2b7290ba39416d8737385e73bb090b3277d1e2