Eliminate accesses to the global `host` option value
The value of the `host` command-line option is accessed via the
`GLOBAL_ARGS` global variable in some places. This causes a problem
because Test Mapping execution code overrides the variant passed to
test runners based on input configuration parameters. This, in turn,
causes the passed in value to be ignored and results in the incorrect
variant of the test being run.

Test: m clean && m atest && atest-dev \
  --test-mapping tools/tradefederation/core
Change-Id: I9bb8a84a1d3901668d0dea1eb76585dfa190d06e