Move tree finding to fileutils, convert test.
git_utils was a weird home for this, and while pytest is better in
almost every way, CI can only handle unittest. Move it to fileutils
and convert the test so it runs in CI.

The move alone isn't very useful, but I'm going to use this more
widely in a follow up as part of trying to remove the requirement that
the CWD is the root of the tree.

Bug: None
Test: None
Change-Id: I2a5b8dfea81b8205d886a745191a5e2ffcce84f1