Autoupdate of 2 submodules
*art updated to 1b126bb39fd6681015ea7c0e6f18b757a846d5c7
Sandbox art test genrules

A number of art tests have custom build steps that depend on other
files not listed in the srcs of the genrule.

Extend test-metadata.json with a `test_build_srcs` property that can be
used to list sources that will be added to the test data genrule.

Bug: 307824623
Test: mm in art/test and ./build/soong/tests/genrule_sandbox_test.py art-run-test-host-data-shard48-tmp and ./art/test.py --host
Change-Id: I2931bcf03216fef43d31e1baf323a41551981358

*build/soong updated to 5eca978092fd348abfc08595db64582867f1239f
Merge "Remove art/test from genrule allowlist" into main